+++ tests/tests_pacasam/test_run_extraction.py
@@ -9,10 +9,22 @@ import glob
 @pytest.mark.timeout(60)
 @pytest.mark.slow
 @pytest.mark.geoportail
-def test_run_extraction_laz(toy_sampling_file):
-    with tempfile.TemporaryDirectory() as tmp_output_path:
+@pytest.mark.parametrize("num_jobs", ["2", "1"])
+# Parallelization needs to be tested first, else there is some issue with laspy.read
+# trying without success to open the cloud that was previously opened successfully in single processing.
+def test_run_extraction_laz(toy_sampling_file, num_jobs):
+    with tempfile.TemporaryDirectory(prefix=f"num_jobs_{num_jobs}_") as tmp_output_path:
         args = parser.parse_args(
-            args=["--sampling_path", toy_sampling_file.name, "--dataset_root_path", tmp_output_path, "--extractor_class", "LAZExtractor"]
+            args=[
+                "--sampling_path",
+                toy_sampling_file.name,
+                "--dataset_root_path",
+                tmp_output_path,
+                "--extractor_class",
+                "LAZExtractor",
+                "--num_jobs",
+                num_jobs,
+            ]
         )
         run_extraction(args)
         created_files = glob.glob(str(args.dataset_root_path / "**/*"))
@@ -22,7 +34,8 @@ def test_run_extraction_laz(toy_sampling_file):
 @pytest.mark.timeout(60)
 @pytest.mark.slow
 @pytest.mark.geoportail
-def test_run_extraction_bd_ortho_today(toy_sampling_file):
+@pytest.mark.parametrize("num_jobs", ["1", "2"])
+def test_run_extraction_bd_ortho_today(toy_sampling_file, num_jobs):
     with tempfile.TemporaryDirectory() as tmp_output_path:
         args = parser.parse_args(
             args=[
@@ -32,6 +45,8 @@ def test_run_extraction_bd_ortho_today(toy_sampling_file):
                 tmp_output_path,
                 "--extractor_class",
                 "BDOrthoTodayExtractor",
+                "--num_jobs",
+                num_jobs,
             ]
         )
         run_extraction(args)
@@ -39,7 +54,8 @@ def test_run_extraction_bd_ortho_today(toy_sampling_file):
         assert len(created_files) == 4
 
 
-def test_run_extraction_bd_ortho_vintage(toy_sampling_file_for_BDOrthoVintageExtractor):
+@pytest.mark.parametrize("num_jobs", ["1", "2"])
+def test_run_extraction_bd_ortho_vintage(toy_sampling_file_for_BDOrthoVintageExtractor, num_jobs):
     os.environ["BD_ORTHO_VINTAGE_VRT_DIR"] = "tests/data/bd_ortho_vintage/"
     os.environ["NUM_JOBS"] = "1"  # default value, here only to give an example
     with tempfile.TemporaryDirectory() as tmp_output_path:
@@ -51,6 +67,8 @@ def test_run_extraction_bd_ortho_vintage(toy_sampling_file_for_BDOrthoVintageExt
                 tmp_output_path,
                 "--extractor_class",
                 "BDOrthoVintageExtractor",
+                "--num_jobs",
+                num_jobs,
             ]
         )
         run_extraction(args)